requirements.txt
black==19.10b0
google-api-python-client==1.8.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
flake8==3.7.9
